Create 0019_RemoveNthNodeFromEndOfList.py
We make sure that the distance between the two pointers is N at all times. So we move the "fast" pointer to by N places at first to ensure that difference between the pointers. due to this, when the fast pointer reaches the last node, the slow pointer is exactly N nodes behind it, *POINTING TO* the node that must be removed.